Bugfixes and minion buffs
- Fixed minions with base critical strike chance not benefitting from it due to an oversight with an earlier big fix
- Added base critical strike chance to Peculiar Sphere and Beetle Idol

+++ Items/Weapons/Summon/Runeterra/RuneterraGauntlets.cs
@@ -34,7 +34,7 @@ namespace tsorcRevamp.Items.Weapons.Summon.Runeterra
         public const int DragonDebuffDuration = 5;
         public const int SuperBurnDuration = 5;
         public const float SuperBurnDmgAmp = 50f;
-        public const float SummonTagCrit = 10f;
+        public const int SummonTagCrit = 10;
         public const float MarkDetonationCritDmgAmp = 2f;
 
         public const float BoostDmgAmp = 25f;
